Round down percentage coverage

+++ usvm-ml-gameserver/src/main/kotlin/org.usvm/JavaMethodRunner.kt
@@ -45,7 +45,7 @@ class JavaMethodRunner(gameMap: GameMap, oracle: Predictor<*>? = null) {
         oracle = oracle
     )
 
-    fun cover(className: String, methodName: String): Pair<List<JcState>, Int> {
+    fun cover(className: String, methodName: String): Pair<List<JcState>, Float> {
         val jcMethod = cp.getJcMethodByName(className, methodName)
 
         val (states, percentageCoverage) = JcMachine(cp, options, interpreterObserver = null).use { machine ->
